docker: Move environment variables to an env file
This way we can have a test environment and a production environment.

To use, you have to copy docker/env.test to .env in this repo's root
directory.

(The reason this is in docker/env.test is so that we don't blow away
changes to the prod env once it's in .env by using git)